+++ src/command-name.ts
@@ -19,12 +19,14 @@ export enum CommandName {
   ProjectSelect = 'ionic-official.projectSelect',
   BuildConfig = 'ionic-official.buildConfig',
   WebConfig = 'ionic-official.webConfig',
+  WebDebugConfig = 'ionic-official.webDebugConfig',
   SelectAction = 'ionic-official.selectAction',
   DebugMode = 'ionic-official.debugMode',
   PluginExplorer = 'ionic-official.pluginExplorer',
   RunMode = 'ionic-official.runMode',
   SelectDevice = 'ionic-official.selectDevice',
   RunIOS = 'ionic-official.run',
+  Debug = 'ionic-official.debug',
   ViewDevServer = 'ionic-official.viewDevServer', // View the dev server window
   hideDevServer = 'ionic-official.hideDevServer', // Hide the dev server window
 }