+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### Version 1.39
 
-- Plugin Explorer added for finding and installing rated Capacitor and Cordova plugins
+- Keyboard shortcut: Alt+R to run your app
+- Keyboard shortcut: Alt+D to run and debug your app
 - Recommend removal of @awesome-cordova-plugin wrappers when plugin is not installed
 - Capacitor 5 upgrade will check for minimum Android Studio version
+- Selection of debug browser is now done through the 3 dots next to "debug"
 - Recommend removal of all deprecated @ionic-native wrappers
 
 ### Version 1.38